Add post date to BlogPostViewModel and display it in BlogPostView
This commit adds a new property, PostDate, to the BlogPostViewModel class that represents the date of a blog post. If the date is null, then PostDateVisible is set to false and the date is not displayed in the view. The commit also updates NewsPanelViewModel to pass the post's create datetime as a DateOnly object when creating a new BlogPostViewModel instance. Finally, this commit modifies BlogPostView.axaml to display PostDate if PostDateVisible is true.