Add post summary to BlogPostViewModel and display it in BlogPostView
This commit adds a new property, PostSummary, to the BlogPostViewModel class. The value of this property is set based on the "summary" field of each blog post. In addition, the BlogPostView has been updated to display this summary when the user hovers over the blog post image. When not hovering over the image, the title of the blog post is displayed instead.