Load blog posts from API (#201)
* Load blog posts from API

* Run dotnet format

* Remove TODO comment for thing already done

* Fix issues reported by Codacy

* Add post date to BlogPostViewModel and display it in BlogPostView

This commit adds a new property, PostDate, to the BlogPostViewModel class that represents the date of a blog post. If the date is null, then PostDateVisible is set to false and the date is not displayed in the view. The commit also updates NewsPanelViewModel to pass the post's create datetime as a DateOnly object when creating a new BlogPostViewModel instance. Finally, this commit modifies BlogPostView.axaml to display PostDate if PostDateVisible is true.

* Add post summary to BlogPostViewModel and display it in BlogPostView

This commit adds a new property, PostSummary, to the BlogPostViewModel class. The value of this property is set based on the "summary" field of each blog post. In addition, the BlogPostView has been updated to display this summary when the user hovers over the blog post image. When not hovering over the image, the title of the blog post is displayed instead.

* Fix Codacy issue